Removed old code. Renamed new code generator.
