+++ lib/loader.js
@@ -1,7 +1,6 @@
 const { Canvg, presets, Parser, SVGElement } = require('canvg');
 const { DOMParser } = require('xmldom');
-// const canvas = require('./code-canvas');
-const canvas=require('./fake-canvas');
+const canvas=require('./code-canvas');
 const CodeGenerator=require('./code-generator');
 const { parseSvgViewBox } = require('./svg-utils');
